requirements-admin.txt
google-auth>=2.45.0
google-auth-oauthlib>=1.2.0,<1.3.0
google-auth-httplib2>=0.3.0
google-api-python-client>=2.157.0
google-cloud-bigquery>=3.26.0
google-cloud-storage>=2.17.0
